created two new files to hold validate function and types
