+++ client/src/Signup/index.ts
@@ -1,0 +1,14 @@
+type ICredentials = {
+  email: string;
+  password: string;
+};
+
+type IErrors = {
+  emailError: { isEmailError: boolean; error: string };
+  passwordError: { isPasswordError: boolean; error: string };
+};
+
+export interface IForm {
+  data: ICredentials;
+  errors: IErrors;
+}
